modification

+++ src/main/java/com/apicollabdev/odk/collabdev/controller/IdeeProjetController.java
@@ -37,7 +37,7 @@ public class IdeeProjetController {
     //  Obtenir une idée par ID
     @GetMapping("/{id}")
     public ResponseEntity<IdeeProjet> getById(@PathVariable Long id) {
-        IdeeProjet idee = ideeProjetService.getById(Math.toIntExact(id));
+        IdeeProjet idee = ideeProjetService.getById((long) Math.toIntExact(id));
         return ResponseEntity.ok(idee);
     }
 
@@ -49,7 +49,7 @@ public class IdeeProjetController {
     // Mettre à jour une idée
     @PutMapping("/{id}")
     public ResponseEntity<IdeeProjet> update(@PathVariable Long id, @RequestBody IdeeProjet updated) {
-        IdeeProjet ideeUpdated = ideeProjetService.updateIdeeProjet(Math.toIntExact(id), updated);
+        IdeeProjet ideeUpdated = ideeProjetService.updateIdeeProjet((long) Math.toIntExact(id), updated);
         return ResponseEntity.ok(ideeUpdated);
     }
